LPD-50639 Rename stable config to smoke config and update references in Playwright configuration
