+++ modules/test/playwright/playwright.config.ts
@@ -119,7 +119,7 @@ import {config as siteNavigationAdminWebConfig} from './tests/site-navigation-ad
 import {config as siteNavigationBreadcrumbWebConfig} from './tests/site-navigation-breadcrumb-web/config';
 import {config as siteNavigationLanguageWebConfig} from './tests/site-navigation-language-web/config';
 import {config as siteNavigationMenuWebConfig} from './tests/site-navigation-menu-web/config';
-import {config as stableConfig} from './tests/stable/config';
+import {config as smokeConfig} from './tests/smoke/config';
 import {config as stagingConfig} from './tests/staging-configuration-web/config';
 import {config as stylebookWebConfig} from './tests/style-book-web/config';
 import {config as templateWebConfig} from './tests/template-web/config';
@@ -250,7 +250,7 @@ export default defineConfig({
 		siteNavigationBreadcrumbWebConfig,
 		siteNavigationLanguageWebConfig,
 		siteNavigationMenuWebConfig,
-		stableConfig,
+		smokeConfig,
 		stagingConfig,
 		stylebookWebConfig,
 		templateWebConfig,
